Analyzing Mechanical System With MATLAB
Position, Velocity and Acceleration Analysis of Offset Slider-Crank Mechanism

Offset slider-crank mechanism is often used for the design of hydraulic driving device in Aviation Machinery. For an offset slider-crank mechanism, the larger its imbalance angle is, the faster its return stroke is, and the higher its work efficiency is in a single-track. Meanwhile, the larger its transmission angle is, the better its motion performance is.